(users): add filter and pagination in users

+++ src/user/user.service.ts
@@ -10,6 +10,8 @@ import { UserRole } from '../user-role/entity/user-role.entity';
 import { UserPlaylist } from '../user-playlist/entity/user-playlist.entity';
 import { Playlist } from '../playlist/entity/playlist.entity';
 import { UpdateUserDto } from './dto';
+import { IDataTable } from '../common/interface';
+import { selectQuery } from '../common/query-builder';
 
 @Injectable()
 export class UserService {
@@ -18,14 +20,21 @@ export class UserService {
     private readonly userRepository: Repository<User>,
   ) {}
 
-  async list() {
+  async list(dataTableOptions: IDataTable) {
     try {
-      const query = this.userRepository
+      const entity = 'user';
+      let query = this.userRepository
         .createQueryBuilder('user')
         .leftJoin(
           UserPlaylist,
           'user_playlist',
           'user_playlist.userId = user.id',
+        )
+        .leftJoinAndMapMany(
+          'user.roles',
+          UserRole,
+          'user_role',
+          'user_role.userId = user.id',
         )
         .leftJoinAndMapMany(
           'user.playlist',
@@ -34,7 +43,30 @@ export class UserService {
           'user_playlist.playlistId = playlist.id',
         );
 
-      return await query.getMany();
+      if (dataTableOptions.filterBy) {
+        if (dataTableOptions.filterBy == 'name') {
+          dataTableOptions.filterValue =
+            dataTableOptions.filterValue.toUpperCase();
+        }
+        query = selectQuery(
+          query,
+          entity,
+          dataTableOptions.filterOperator,
+          dataTableOptions.filterBy,
+          dataTableOptions.filterValue,
+        );
+      }
+
+      const skip = dataTableOptions.pageSize * dataTableOptions.pageIndex || 0;
+
+      return await query
+        .skip(skip)
+        .take(dataTableOptions.pageSize)
+        .orderBy(
+          `${entity}.${dataTableOptions.sortBy}`,
+          dataTableOptions.sortOrder,
+        )
+        .getManyAndCount();
     } catch (error) {
       throw new InternalServerErrorException();
     }
@@ -127,10 +159,16 @@ export class UserService {
     }
   }
 
-  async remove(id: string) {
+  async remove(currentId: string, id: string) {
     try {
+      if (currentId === id) {
+        throw new BadRequestException('cannot delete this user');
+      }
       await this.userRepository.delete(id);
     } catch (error) {
+      if (error instanceof BadRequestException) {
+        throw new BadRequestException(error.message);
+      }
       throw new InternalServerErrorException();
     }
   }
